Move resolver information formatting to the plugin.
